refactor: 비대면 진료실에서 진단서 작성방식 appointment에서 가져와서 구현

+++ src/main/java/com/codelion/animalcare/domain/diagnosis/dto/FindOneDiagnosis.java
@@ -34,9 +34,9 @@ public class FindOneDiagnosis {
     // 동물명
     private String animalName;
     // 성별
-    private short animalGenderId;
+    private int animalGenderId;
     // 연령
-    private short animalAge;
+    private int animalAge;
     // 모색
     private String animalCoatColor;
     // 특징
